+++ CHANGELOG.rst
@@ -5,4 +5,6 @@ Releases
 Notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`__.
 
-No releases yet!
+Latest
+------
+* Added functionality to create a CSV file containing the latest directory listing of files, filtered by instrument. The output file is named "latest_dirlist.csv" and is generated during the calibration process.